+++ opennms-services/src/main/java/org/opennms/netmgt/eventd/processor/JdbcEventWriter.java
@@ -182,8 +182,8 @@ public final class JdbcEventWriter extends AbstractJdbcPersister implements Even
             insStmt.setString(2, Constants.format(event.getUei(), EVENT_UEI_FIELD_SIZE));
 
             // nodeID
-            final int nodeid = (int) event.getNodeid();
-            set(insStmt, 3, event.hasNodeid() ? nodeid : -1);
+            final Long nodeid = event.getNodeid();
+            set(insStmt, 3, event.hasNodeid() ? nodeid.intValue() : -1);
 
             // eventTime
             insStmt.setTimestamp(4, getEventTime(event));
@@ -421,7 +421,7 @@ public final class JdbcEventWriter extends AbstractJdbcPersister implements Even
         }
         
         try {
-            return getHostName((int) event.getNodeid(), event.getHost(), connection);
+            return getHostName(event.getNodeid().intValue(), event.getHost(), connection);
         } catch (final Throwable t) {
             LogUtils.warnf(this, t, "Error converting host IP \"%s\" to a hostname, storing the IP.", event.getHost());
             return event.getHost();
